requirements.txt
ultralytics>=8.0.0
requests
opencv-python
openai
google-generativeai
anthropic
numpy
pillow
python-dotenv

